move injectPolyfillsIntoEntry to haul-core
