+++ packages/haul-preset-0.59/package.json
@@ -16,7 +16,6 @@
   "peerDependencies": {
     "@haul-bundler/cli": "1.0.0-rc.15",
     "@haul-bundler/core": "1.0.0-rc.15",
-    "@haul-bundler/core-legacy": "1.0.0-rc.15",
     "webpack": "^4.30.0"
   },
   "dependencies": {
@@ -28,7 +27,6 @@
   "devDependencies": {
     "@haul-bundler/cli": "^0.3.0",
     "@haul-bundler/core": "^0.3.0",
-    "@haul-bundler/core-legacy": "^0.3.0",
     "@types/case-sensitive-paths-webpack-plugin": "^2.1.2",
     "@types/is-ci": "^1.1.0",
     "@types/terser-webpack-plugin": "^1.2.1",
